+++ src/name/NicolasCoutureGrenier/Maths/FuzzyLogic/Valuators/CombinationEntropy.java
@@ -21,7 +21,7 @@ public class CombinationEntropy implements Function<Combination, FuzzyVariable> 
     double max = Double.MIN_VALUE;
 
     if (!minmax.containsKey(n)) {
-      int k = Numbers.reverseTriangularNumber(n);
+      long k = Numbers.reverseTriangularNumber(n);
       max = Math.log(k);
       minmax.put(n, Interval.makeClosedInterval(0.0, max+0.0000000000000005));
     } else {
